refactor: extract ui contract v2 projection helpers
